Fix JSONB array access for variant stock reservation

+++ src/app/api/orders/route.ts
@@ -174,14 +174,14 @@ export async function POST(req: NextRequest) {
       ok = await db
         .update(products)
         .set({
-          variants: sql`jsonb_set(${products.variants}::jsonb, ARRAY[${vi}::text, 'stock'], to_jsonb((${products.variants}::jsonb->${vi}->>'stock')::integer - ${item.quantity}))::json`,
+          variants: sql`jsonb_set(${products.variants}::jsonb, ARRAY[${vi}::text, 'stock'], to_jsonb((${products.variants}::jsonb->(${vi})::int->>'stock')::integer - ${item.quantity}))::json`,
           // Mirror first-variant decrement to flat stock so product cards stay accurate
           ...(vi === 0 ? { stock: sql`GREATEST(${products.stock} - ${item.quantity}, 0)` } : {}),
           orderCount: sql`${products.orderCount} + ${item.quantity}`,
         })
         .where(and(
           eq(products.id, item.productId),
-          sql`(${products.variants}::jsonb->${vi}->>'stock')::integer >= ${item.quantity}`
+          sql`(${products.variants}::jsonb->(${vi})::int->>'stock')::integer >= ${item.quantity}`
         ))
         .returning({ id: products.id });
     } else {
